requirements.txt
django
gunicorn
django-rest-framework
django-bootstrap3==12.0.3
